refactor(incident): split monitor module
